[FEAT] : ⚙️: 검색 아이템 구분
검색 아이템 구분하게됨 및 비디오 검색 수정
