+++ app/src/main/java/com/example/assignmnet_img/bookmark/BookmarkModel.kt
@@ -5,6 +5,7 @@ import com.example.assignmnet_img.search.dataclass.SearchModel
 data class BookmarkModel(
     val id: Long = -1,
     val imageUrl: String,
+    val label:String,
     val displaySiteName: String,
     val datetime: String,
     val isBookmark: Boolean = false
@@ -14,7 +15,8 @@ fun BookmarkModel.toSearchModel(): SearchModel {
     return SearchModel(
         id = id,
         Url = imageUrl,
-        displaySiteName = displaySiteName,
+        label = label,
+        title = displaySiteName,
         datetime = datetime,
         isBookmark = isBookmark
     )
